fix: rename package directory to not contain underscores
