+++ pkg/metadataquery/query.go
