+++ pkg/metadataquery/series.go
